GSC/pyproject.toml
[project]
name = "gsc"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.13.3",
    "black>=25.1.0",
    "firebase-admin>=6.7.0",
    "flask>=3.1.0",
    "google-cloud-firestore>=2.20.1",
    "google-cloud-language>=2.17.1",
    "google-cloud-storage>=3.1.0",
    "google-cloud-vision>=3.10.1",
    "google-generativeai>=0.8.4",
    "kafka-python>=2.1.5",
    "pyspark>=3.5.5",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "scikit-learn>=1.6.1",
]
